Splits TArrayIteratorTrait into TArrayCopyIteratorTrait
